make stow not stow stuff for wrong hostname

+++ stow-confs.sh
@@ -10,9 +10,12 @@ stow -R . -t ~/docs/scripts/ -v
 cd ..
 cd others || exit
 stow -R . -t ~/.local/share/ -v 
-cd ..
-cd configs/lisgd-0.4.0 || exit
-make
-killall lisgd
-cp lisgd ~/.local/bin/lisgd
-lisgd &
+HOSTNAME=$(cat /etc/hostname)
+if [[ $HOSTNAME = "weasel" ]] then
+   cd ..
+   cd configs/lisgd-0.4.0 || exit
+   make
+   killall lisgd
+   cp lisgd ~/.local/bin/lisgd
+   lisgd &
+fi
